% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dtw_heatmap.R
\name{plot.scRhythmDTWHeatmap}
\alias{plot.scRhythmDTWHeatmap}
\title{Plot a scRhythmDTWHeatmap object}
\usage{
\method{plot}{scRhythmDTWHeatmap}(
  x,
  merge_legends = TRUE,
  heatmap_legend_side = "right",
  annotation_legend_side = "right",
  padding = grid::unit(c(6, 4, 4, 4), "mm"),
  ...
)
}
\arguments{
\item{x}{A \code{scRhythmDTWHeatmap} object returned by
\code{plot_dtw_heatmap()}.}

\item{merge_legends}{A logical value indicating whether heatmap and
annotation legends should be merged. Defaults to \code{TRUE}.}

\item{heatmap_legend_side}{A character value specifying the heatmap legend
side. Defaults to \code{"right"}.}

\item{annotation_legend_side}{A character value specifying the annotation
legend side. Defaults to \code{"right"}.}

\item{padding}{A \code{\link[grid]{unit}} object specifying padding passed to
\code{ComplexHeatmap::draw()}. Defaults to
\code{grid::unit(c(6, 4, 4, 4), "mm")}.}

\item{...}{Additional arguments passed to \code{ComplexHeatmap::draw()}.}
}
\value{
Invisibly returns the drawn ComplexHeatmap object.
}
\description{
This S3 method draws a \code{scRhythmDTWHeatmap} object returned by
\code{plot_dtw_heatmap()}.
}
