% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_ovr_rhythm.R
\name{plot.scRhythmOVRHeatmap}
\alias{plot.scRhythmOVRHeatmap}
\title{Print a scRhythmOVRHeatmap object}
\usage{
\method{plot}{scRhythmOVRHeatmap}(x, ...)
}
\arguments{
\item{x}{A \code{scRhythmOVRHeatmap} object.}

\item{...}{Additional arguments.}
}
\value{
Invisibly returns \code{x}.
}
\description{
This S3 method prints a concise summary of a
\code{scRhythmOVRHeatmap} object returned by
\code{plot_ovr_heatmap()}.
}
