% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dtw_heatmap.R
\name{print.scRhythmDTWHeatmap}
\alias{print.scRhythmDTWHeatmap}
\title{Print a scRhythmDTWHeatmap object}
\usage{
\method{print}{scRhythmDTWHeatmap}(x, ...)
}
\arguments{
\item{x}{A \code{scRhythmDTWHeatmap} object.}

\item{...}{Additional arguments.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a scRhythmDTWHeatmap object
}
