% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_dtw_similarity.R
\name{print.scRhythmDTWSimilarity}
\alias{print.scRhythmDTWSimilarity}
\title{Print a scRhythmDTWSimilarity object}
\usage{
\method{print}{scRhythmDTWSimilarity}(x, ...)
}
\arguments{
\item{x}{A \code{scRhythmDTWSimilarity} object returned by
\code{run_dtw_similarity()}.}

\item{...}{Additional arguments.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a scRhythmDTWSimilarity object
}
