% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_module.R
\name{print.scRhythmModuleHeatmap}
\alias{print.scRhythmModuleHeatmap}
\title{Print a scRhythmModuleHeatmap object}
\usage{
\method{print}{scRhythmModuleHeatmap}(x, ...)
}
\arguments{
\item{x}{A \code{scRhythmModuleHeatmap} object.}

\item{...}{Additional arguments.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a scRhythmModuleHeatmap object
}
