% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_ovr_rhythm.R
\name{print.scRhythmOVRHeatmap}
\alias{print.scRhythmOVRHeatmap}
\title{Plot a scRhythmOVRHeatmap object}
\usage{
\method{print}{scRhythmOVRHeatmap}(x, ...)
}
\arguments{
\item{x}{A \code{scRhythmOVRHeatmap} object.}

\item{...}{Additional arguments passed to \code{ComplexHeatmap::draw()}.}
}
\value{
Invisibly returns the drawn ComplexHeatmap object.
}
\description{
This S3 method draws the ComplexHeatmap object stored in a
\code{scRhythmOVRHeatmap} object returned by \code{plot_ovr_heatmap()}.
}
