% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_phase_rho.R
\name{print.scRhythmPhaseRho}
\alias{print.scRhythmPhaseRho}
\title{Print a scRhythmPhaseRho object}
\usage{
\method{print}{scRhythmPhaseRho}(x, ...)
}
\arguments{
\item{x}{A \code{scRhythmPhaseRho} object.}

\item{...}{Additional arguments. Currently unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
This S3 method prints a concise summary of a
\code{scRhythmPhaseRho} object returned by \code{run_phase_rho()}.
}
