% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_theme.R
\name{theme_srb}
\alias{theme_srb}
\title{scRhythmBayes ggplot theme}
\usage{
theme_srb(aspect.ratio = NULL, base_size = 12, base_family = "", ...)
}
\arguments{
\item{aspect.ratio}{A numeric value specifying the panel aspect ratio.
Defaults to \code{NULL}.}

\item{base_size}{A numeric value specifying the base font size. Defaults to
\code{12}.}

\item{base_family}{A character vector specifying the base font family.
Defaults to \code{""}.}

\item{...}{Additional arguments passed to \code{\link[ggplot2]{theme}}.}
}
\value{
A ggplot2 theme object.
}
\description{
This function provides the default ggplot2 theme used by scRhythmBayes
plotting functions.
}
\examples{
\dontrun{
library(ggplot2)

p <- ggplot(mtcars, aes(x = wt, y = mpg)) +
  geom_point()

p + theme_srb()
}

}
