% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_theme.R
\name{theme_srb_void}
\alias{theme_srb_void}
\title{scRhythmBayes void ggplot theme}
\usage{
theme_srb_void(base_size = 12, base_family = "", ...)
}
\arguments{
\item{base_size}{A numeric value specifying the base font size. Defaults to
\code{12}.}

\item{base_family}{A character vector specifying the base font family.
Defaults to \code{""}.}

\item{...}{Additional arguments passed to \code{\link[ggplot2]{theme}}.}
}
\value{
A ggplot2 theme object.
}
\description{
This function provides a blank ggplot2 theme used by circular or
annotation-only scRhythmBayes plotting functions.
}
\examples{
\dontrun{
library(ggplot2)

p <- ggplot(mtcars, aes(x = wt, y = mpg)) +
  geom_point()

p + theme_srb_void()
}

}
